requirements.txt
asgiref==3.5.2
colorama==0.4.5
colorlog==6.6.0
Django==4.0.5
django-debug-toolbar==3.5.0
django-dynamic-preferences==1.13.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-mptt-admin==2.2.0
persisting-theory==1.0
Pillow==9.2.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
